added bucket search. untested

+++ src/Models/Storage/M_BucketRow.cs
@@ -4,6 +4,6 @@ namespace Agent.Models.Storage;
 public class BucketRow
 {
     public required long id { get; set; }
-    public required double[] vector { get; set; }
+    public required float[] vector { get; set; }
     public required byte[] chunk { get; set; }
 }
